#79 [RENAME] CalculateTimeFromString 파일 rename
