+++ presentation/src/main/java/com/sopt/presentation/appointment/appointmentConfirm/AppointmentConfirmRoute.kt
@@ -38,7 +38,7 @@ import com.sopt.core.designsystem.theme.NoostakAndroidTheme
 import com.sopt.core.designsystem.theme.NoostakTheme
 import com.sopt.core.extension.showIf
 import com.sopt.core.state.UiState
-import com.sopt.core.util.CalculateTime
+import com.sopt.core.util.time.CalculateTimeFromString
 import com.sopt.core.util.RearrangeList
 import com.sopt.domain.entity.AppointmentDetailEntity
 import com.sopt.domain.entity.IdentityEntity
@@ -130,7 +130,7 @@ fun AppointmentConfirmScreen(
     onConfirmButtonClick: (Long) -> Unit,
     data: AppointmentDetailEntity
 ) {
-    val calculateTime = CalculateTime()
+    val calculateTime = CalculateTimeFromString()
     val date = calculateTime.extractDateWithKorean(data.date)
     val dayOfWeek = calculateTime.extractDayOfWeekWithBraces(data.date)
     val startHour = calculateTime.extractHourWithZero(data.startTime)
